fix(node-runtime-worker-thread): Propagate child process startup errors for thread runtime (#721)
* fix(node-runtime-worker-thread): Handle errors that can happen on worker or child process start

* fix(compass-shell): Change runtime resolution to fix running compass-shell locally

* fix(node-runtime-worker-thread): Set encoding for stderr to avoid broken characters when collecting stream chunks

+++ packages/node-runtime-worker-thread/src/index.ts
@@ -10,7 +10,7 @@ import {
 } from '@mongosh/browser-runtime-core';
 import { MongoshBus } from '@mongosh/types';
 import path from 'path';
-import { EventEmitter } from 'events';
+import { EventEmitter, once } from 'events';
 import { kill } from './spawn-child-from-source';
 import { Caller, createCaller, cancel } from './rpc';
 import { ChildProcessEvaluationListener } from './child-process-evaluation-listener';
@@ -19,6 +19,31 @@ import { deserializeEvaluationResult } from './serializer';
 import { ChildProcessMongoshBus } from './child-process-mongosh-bus';
 
 type ChildProcessRuntime = Caller<WorkerThreadWorkerRuntime>;
+
+function parseStderrToError(str: string): Error | null {
+  const [, errorMessageWithStack] = str
+    .split(/^\s*\^\s*$/m)
+    .map((part) => part.trim());
+
+  if (errorMessageWithStack) {
+    const e = new Error();
+    const errorHeader =
+      errorMessageWithStack.substring(
+        0,
+        errorMessageWithStack.search(/^\s*at/m)
+      ) || errorMessageWithStack;
+
+    const [name, ...message] = errorHeader.split(': ');
+
+    e.name = name;
+    e.message = message.join(': ').trim();
+    e.stack = errorMessageWithStack;
+
+    return e;
+  }
+
+  return null;
+}
 
 class WorkerRuntime implements Runtime {
   private initOptions: {
@@ -42,6 +67,11 @@ class WorkerRuntime implements Runtime {
 
   private initWorkerPromise: Promise<void>;
 
+  private childProcessProxySrcPath: string =
+    process.env
+      .CHILD_PROCESS_PROXY_SRC_PATH_DO_NOT_USE_THIS_EXCEPT_FOR_TESTING ||
+    path.resolve(__dirname, 'child-process-proxy.js');
+
   constructor(
     uri: string,
     driverOptions: MongoClientOptions = {},
@@ -57,15 +87,48 @@ class WorkerRuntime implements Runtime {
   private async initWorker() {
     const { uri, driverOptions, cliOptions, spawnOptions } = this.initOptions;
 
-    const childProcessProxySrcPath = path.resolve(
-      __dirname,
-      'child-process-proxy.js'
+    this.childProcess = spawn(
+      process.execPath,
+      [this.childProcessProxySrcPath],
+      {
+        stdio: ['inherit', 'inherit', 'pipe', 'ipc'],
+        ...spawnOptions
+      }
     );
 
-    this.childProcess = spawn(process.execPath, [childProcessProxySrcPath], {
-      stdio: ['inherit', 'inherit', 'inherit', 'ipc'],
-      ...spawnOptions
+    const waitForReadyMessage = async() => {
+      let msg: string;
+      while (([msg] = await once(this.childProcess, 'message'))) {
+        if (msg === 'ready') return;
+      }
+    };
+
+    let spawnError = '';
+
+    // eslint-disable-next-line chai-friendly/no-unused-expressions
+    this.childProcess?.stderr?.setEncoding('utf8')?.on('data', (chunk) => {
+      spawnError += chunk;
     });
+
+    const waitForError = async() => {
+      const [exitCode] = await once(this.childProcess, 'exit');
+
+      if (exitCode) {
+        let error = parseStderrToError(spawnError);
+
+        if (error) {
+          error.message = `Child process failed to start with the following error: ${error.message}`;
+        } else {
+          error = new Error(
+            `Worker runtime failed to start: child process exited with code ${exitCode}`
+          );
+        }
+
+        throw error;
+      }
+    };
+
+    await Promise.race([waitForReadyMessage(), waitForError()]);
 
     // We expect the amount of listeners to be more than the default value of 10
     // but probably not more than ~25 (all exposed methods on
@@ -122,11 +185,26 @@ class WorkerRuntime implements Runtime {
   }
 
   async terminate() {
-    await this.initWorkerPromise;
+    try {
+      await this.initWorkerPromise;
+    } catch {
+      // In case child process encountered an error during init we still want
+      // to clean up whatever possible
+    }
+
     await kill(this.childProcess);
-    this.childProcessRuntime[cancel]();
-    this.childProcessEvaluationListener.terminate();
-    this.childProcessMongoshBus.terminate();
+
+    if (this.childProcessRuntime) {
+      this.childProcessRuntime[cancel]();
+    }
+
+    if (this.childProcessEvaluationListener) {
+      this.childProcessEvaluationListener.terminate();
+    }
+
+    if (this.childProcessMongoshBus) {
+      this.childProcessMongoshBus.terminate();
+    }
   }
 
   async interrupt() {
